backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.119.0",
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.0",
    "langgraph>=0.6.10",
    "psycopg2-binary>=2.9.11",
    "pyclerk>=0.0.1",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.37.0",
]
